+++ src/main/java/com/example/demo/controller/BookController.java
@@ -1,10 +1,15 @@
 package com.example.demo.controller;
 
+import com.example.demo.domain.ErrorResponse;
 import com.example.demo.domain.dto.BookDto;
+import com.example.demo.domain.entity.AuthorEntity;
 import com.example.demo.domain.entity.BookEntity;
 import com.example.demo.mapper.Mapper;
+import com.example.demo.service.AuthorService;
 import com.example.demo.service.BookService;
 import lombok.AllArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
@@ -19,18 +24,21 @@ import java.util.Optional;
 //@RequestMapping("/api")
 public class BookController {
 
+    private static final Logger log = LoggerFactory.getLogger(BookController.class);
+
     private Mapper<BookEntity, BookDto> bookMapper;
 
     private BookService bookService;
+    private AuthorService authorService;
 
     @PutMapping("/books/{isbn}")
     public ResponseEntity<BookDto> createUpdateBook(
-            @PathVariable String isbn,
+            @PathVariable("isbn") String isbnId,
             @RequestBody BookDto bookDto
     ) {
         BookEntity bookEntity = bookMapper.mapFrom(bookDto);
-        boolean bookExists = bookService.isExists(isbn);
-        BookEntity savedBookEntity = bookService.createUpdateBook(isbn, bookEntity);
+        boolean bookExists = bookService.isExists(isbnId);
+        BookEntity savedBookEntity = bookService.createUpdateBook(isbnId, bookEntity);
         BookDto savedUpdatedBookDto = bookMapper.mapTo(savedBookEntity);
 
         if (bookExists) {
@@ -38,6 +46,50 @@ public class BookController {
         } else {
             return new ResponseEntity<>(savedUpdatedBookDto, HttpStatus.CREATED);
         }
+    }
+
+    @PutMapping("/books/{isbn}/{authorId}")
+    public ResponseEntity<?> createUpdateBookWithAuthorId(
+            @PathVariable String isbn,
+            @PathVariable String authorId,
+            @RequestBody BookDto bookDto
+    ) {
+        BookEntity bookEntity = bookMapper.mapFrom(bookDto);
+        boolean bookExists = bookService.isExists(isbn);
+        long id;
+        try {
+            id = Long.parseLong(authorId);
+        } catch (NumberFormatException e) {
+            log.error("Error parsing author ID", e);
+            return ResponseEntity
+                    .badRequest()
+                    .body(new ErrorResponse("Invalid author ID format"));
+        }
+
+        Optional<AuthorEntity> author = authorService.findOne(id);
+
+        BookEntity savedBookEntity;
+        if (author.isPresent()) {
+            bookEntity.setAuthor(author.get());
+            savedBookEntity = bookService.createUpdateBook(isbn, bookEntity);
+        } else {
+            bookEntity.setAuthor(null);
+            return ResponseEntity
+                    .badRequest()
+                    .body( new ErrorResponse("Author not found") );
+        }
+
+        if (savedBookEntity == null) {
+            return ResponseEntity
+                    .badRequest()
+                    .body(new ErrorResponse("Book could not be saved, error createUpdateBook(isbn, bookEntity)"));
+        }
+        BookDto savedUpdatedBookDto = bookMapper.mapTo(savedBookEntity);
+
+        return new ResponseEntity<>(
+                savedUpdatedBookDto,
+                bookExists ? HttpStatus.OK : HttpStatus.CREATED
+        );
     }
 
     @PatchMapping("/books/{isbn}")
@@ -77,9 +129,13 @@ public class BookController {
     }
 
     @DeleteMapping(path = "/books/{isbn}")
-    public ResponseEntity deleteBook(@PathVariable String isbn) {
-        bookService.delete(isbn);
-        return new ResponseEntity(HttpStatus.NO_CONTENT);
+    public ResponseEntity<String> deleteBook(@PathVariable String isbn) {
+        if (isbn != null && bookService.isExists(isbn)) {
+            bookService.delete(isbn);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>("Wrong isbn ID", HttpStatus.BAD_REQUEST);
+        }
     }
 
 }
